Extend monioring URL detection to allow custom prefixes
In Gerrit Code Review auth requests are prepended with "/a" prefix.
Extend the comparisson for monitoring URI detection to allow custom
prefixes and only require that the URL is ending with pre-defined
monitoring path.

This is need to fix Prometheus scraping from Gerrit, while allowing
the authentication machinery to work as expected:

  curl -u user http://gerrit/a/monitoring?format=prometheus