Extend monioring URL detection for authenticated use case
In Gerrit Code Review auth requests are prepended with "/a" prefix.
Extend the comparisson for monitoring URI detection to support this
use case. Add new configuration option: monitoring-path-authenticated
and check the URL for the value of this option if set. Given that the
new configuration option is unset per default, the performance overhead
is near to zero as the only added condition is the non null check for
sites that don't utilize that option.

Considered alterative is to replace the full match check with suffix
check only (endsWith instead of equals) but this could be dangerous,
in case the applications itself expose endpoints with "/*/monitoring
suffix.

This is need to fix Prometheus scraping from Gerrit, while allowing
the authentication machinery to work as expected:

  curl -u user http://gerrit/a/monitoring?format=prometheus