support for HTTP headers in collector server
When using collector server to pull data from monitored clients, HTTP
headers need to be supported. During registration with the collector
server, monitored clients can now provide a map of headers. This will
be used by collector server when calling the monitored clients. This
scenario is useful for example in cloud foundry, when there are many
instances of an application behind a load balancer and we want to
monitor each application separately. For such scenarios, cloud foundry
uses a header X-CF-APP-INSTANCE to identify the specific instance.